pyproject.toml
[tool.poetry]
name = "template_system"
version = "0.1.0"
description = ""
authors = ["Erick Duarte <erickod@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.75.0"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
python-multipart = "^0.0.5"
python-docx = "^0.8.11"
setuptools = "^60.10.0"
httpx = "^0.22.0"
dynaconf = "^3.1.7"

[tool.poetry.dev-dependencies]
pytest = "^7.1.0"
mypy = "^0.941"
black = "^22.1.0"
flake8 = "^4.0.1"
setuptools = "^60.10.0"
isort = "^5.10.1"
pytest-cov = "^3.0.0"
requests = "^2.27.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
